Correct computation of fee paid by chain

+++ WalletWasabi/Models/CpfpInfo.cs
@@ -4,4 +4,4 @@ using NBitcoin;
 namespace WalletWasabi.Models;
 
 public record AncestorCpfpInfo(uint256 TxId, long Fee, long Weight);
-public record CpfpInfo(List<AncestorCpfpInfo> Ancestors, double EffectiveFeePerVSize, double AdjustedVSize);
+public record CpfpInfo(List<AncestorCpfpInfo> Ancestors, decimal EffectiveFeePerVSize, decimal AdjustedVSize);
